Extract hooks for quick receipt, seller profile and user settings
